long coercsion

+++ utils/sipwitch.cpp
@@ -591,11 +591,11 @@ static void dumpstats(char **argv)
         else if(!map.lastcall)
             printf("%s -\n", text);
         else if(now - map.lastcall > (3600l * 99l))
-            printf("%s %ld%c\n", text, (now - map.lastcall) / (3600l * 24l), 'd');
+            printf("%s %ld%c\n", text, (long)((now - map.lastcall) / (3600l * 24l)), 'd');
         else if(now - map.lastcall > (60l * 120l))
-            printf("%s %ld%c\n", text, (now - map.lastcall) / 3600l, 'h');
+            printf("%s %ld%c\n", text, (long)((now - map.lastcall) / 3600l), 'h');
         else if(now - map.lastcall > 120l)
-            printf("%s %ld%c\n", text, (now - map.lastcall) / 60l, 'm');
+            printf("%s %ld%c\n", text, (long)((now - map.lastcall) / 60l), 'm');
         else
             printf("%s %ld%c\n", text, (long)(now - map.lastcall), 's');
     }
